+++ docs/src/submodules/Bridges/list_of_bridges.md
@@ -47,7 +47,6 @@ Bridges.Constraint.NormNuclearBridge
 Bridges.Constraint.SquareBridge
 Bridges.Constraint.RootDetBridge
 Bridges.Constraint.LogDetBridge
-Bridges.Constraint.DetSquareBridge
 Bridges.Constraint.IndicatorActiveOnFalseBridge
 Bridges.Constraint.IndicatorSOS1Bridge
 Bridges.Constraint.SemiToBinaryBridge
