+++ src/Bridges/Bridges.jl
@@ -245,19 +245,11 @@ julia> MOI.Bridges.runtests(
        )
 ```
 """
-function runtests(
-    Bridge::Type{<:AbstractBridge},
-    input::String,
-    output::String;
-    debug::Bool = false,
-)
+function runtests(Bridge::Type{<:AbstractBridge}, input::String, output::String)
     # Load model and bridge it
     inner = MOI.Utilities.UniversalFallback(MOI.Utilities.Model{Float64}())
     model = _bridged_model(Bridge, inner)
     MOI.Utilities.loadfromstring!(model, input)
-    if debug
-        print(inner)
-    end
     # Load a non-bridged input model, and check that getters are the same.
     test = MOI.Utilities.UniversalFallback(MOI.Utilities.Model{Float64}())
     MOI.Utilities.loadfromstring!(test, input)
